nodes/firmware/custom-modules/flashdb_store_service: add lock for service operations
This seems to fix the flashdb corruption (CRC errors / duplicate entries / dropped entries etc.) which is probably triggered by different threads accessing the service.
The locking could probably be done in a less restrictive way but the root cause (which specific sequence of operations causes the issue) could not be isolated yet.